Add customizable track selection
- Simplify standard track selections
  - Separate implementation file
  - reuse previous functions
  - name ITS2 cuts with Run3

- Add Run3 track requirement for SSD tracks
