+++ Common/TableProducer/trackselection.cxx
@@ -38,6 +38,10 @@ struct TrackSelectionTask {
   // FIXME: this will be removed once we can get this via meta data
   Configurable<bool> isRun3{"isRun3", false, "temp option to enable run3 mode"};
   Configurable<int> itsMatching{"itsMatching", 0, "condition for ITS matching (0: Run2 SPD kAny, 1: Run3ITSibAny, 2: Run3ITSallAny, 3: Run3ITSall7Layers)"};
+  Configurable<float> ptMin{"ptMin", 0.1f, "Lower cut on pt for the track selected"};
+  Configurable<float> ptMax{"ptMax", 1e10f, "Upper cut on pt for the track selected"};
+  Configurable<float> etaMin{"etaMin", -0.8, "Lower cut on eta for the track selected"};
+  Configurable<float> etaMax{"etaMax", 0.8, "Upper cut on eta for the track selected"};
 
   Produces<aod::TrackSelection> filterTable;
 
@@ -54,30 +58,35 @@ struct TrackSelectionTask {
       case 1:
         // Run 3 kAny on 3 IB layers of ITS
         if (isRun3) {
-          globalTracks = getGlobalTrackSelectionITSMatch(TrackSelection::GlobalTrackRun3ITSMatching::Run3ITSibAny);
+          globalTracks = getGlobalTrackSelectionRun3ITSMatch(TrackSelection::GlobalTrackRun3ITSMatching::Run3ITSibAny);
         }
         break;
       case 2:
         // Run 3 kAny on all 7 layers of ITS
         if (isRun3) {
-          globalTracks = getGlobalTrackSelectionITSMatch(TrackSelection::GlobalTrackRun3ITSMatching::Run3ITSallAny);
+          globalTracks = getGlobalTrackSelectionRun3ITSMatch(TrackSelection::GlobalTrackRun3ITSMatching::Run3ITSallAny);
         }
         break;
       case 3:
         // Run 3 kAll on all 7 layers of ITS
         if (isRun3) {
-          globalTracks = getGlobalTrackSelectionITSMatch(TrackSelection::GlobalTrackRun3ITSMatching::Run3ITSall7Layers);
+          globalTracks = getGlobalTrackSelectionRun3ITSMatch(TrackSelection::GlobalTrackRun3ITSMatching::Run3ITSall7Layers);
         }
         break;
-
       default:
         LOG(fatal) << "TrackSelectionTask with undefined cuts. Fix it!";
         break;
     }
-    globalTracksSDD = getGlobalTrackSelectionSDD();
+    globalTracks.SetPtRange(ptMin, ptMax);
+    globalTracks.SetEtaRange(etaMin, etaMax);
 
+    // Extra requirement on the ITS -> Run 2: asking for 1 hit SDD and no hit in SPD
+    globalTracksSDD = getGlobalTrackSelectionSDD();
+    globalTracksSDD.SetPtRange(ptMin, ptMax);
+    globalTracksSDD.SetEtaRange(etaMin, etaMax);
     if (isRun3) {
-      globalTracks.SetTrackType(o2::aod::track::TrackTypeEnum::Track);
+      globalTracks.SetTrackType(o2::aod::track::TrackTypeEnum::Track);    // Requiring that this is a Run 3 track
+      globalTracksSDD.SetTrackType(o2::aod::track::TrackTypeEnum::Track); // Requiring that this is a Run 3 track
     }
   }
